+++ hahomematic/central/__init__.py
@@ -1140,7 +1140,12 @@ class CentralUnit:
 
     def get_sysvar_entity(self, name: str) -> GenericSystemVariable | None:
         """Return the sysvar entity."""
-        return self._sysvar_entities.get(name)
+        if sysvar := self._sysvar_entities.get(name):
+            return sysvar
+        for sysvar in self._sysvar_entities.values():
+            if sysvar.name == name:
+                return sysvar
+        return None
 
     def get_program_button(self, pid: str) -> HmProgramButton | None:
         """Return the program button."""