+++ hahomematic/central/__init__.py
@@ -6,7 +6,7 @@ This is the python representation of a CCU.
 from __future__ import annotations
 
 import asyncio
-from collections.abc import Awaitable, Callable, Coroutine, Set
+from collections.abc import Awaitable, Callable, Coroutine, Mapping, Set
 from concurrent.futures._base import CancelledError
 from datetime import datetime
 import logging
@@ -58,7 +58,7 @@ from hahomematic.performance import measure_execution_time
 from hahomematic.platforms import create_entities_and_append_to_device
 from hahomematic.platforms.custom.entity import CustomEntity
 from hahomematic.platforms.device import HmDevice
-from hahomematic.platforms.entity import BaseEntity
+from hahomematic.platforms.entity import BaseEntity, CallbackEntity
 from hahomematic.platforms.event import GenericEvent
 from hahomematic.platforms.generic.entity import GenericEntity, WrapperEntity
 from hahomematic.platforms.hub import Hub
@@ -142,8 +142,6 @@ class CentralUnit:
         self._entity_event_subscriptions: Final[dict[tuple[str, str], Any]] = {}
         # {unique_identifier, entity}
         self._entities: Final[dict[str, BaseEntity]] = {}
-        # {unique_identifier, update_entity}
-        self._firmware_update_entities: Final[dict[str, HmUpdate]] = {}
         # {device_address, device}
         self._devices: Final[dict[str, HmDevice]] = {}
         # {sysvar_name, sysvar_entity}
@@ -151,7 +149,7 @@ class CentralUnit:
         # {sysvar_name, program_button}U
         self._program_buttons: Final[dict[str, HmProgramButton]] = {}
         # {unique_identifier}
-        self._subscribed_entity_unique_identifiers: Final[set[str]] = set()
+        self._subscribed_entity_unique_identifiers: Final[dict[str, str]] = {}
         # store last event received datetime by interface
         self.last_events: Final[dict[str, datetime]] = {}
         # Signature: (name, *args)
@@ -298,14 +296,16 @@ class CentralUnit:
         """Return the unique identifiers of subscribed entities."""
         return tuple(self._subscribed_entity_unique_identifiers)
 
-    def add_subscribed_entity_unique_identifier(self, unique_identifier: str) -> None:
+    def add_subscribed_entity_unique_identifier(
+        self, unique_identifier: str, custom_identifier: str
+    ) -> None:
         """Add new program button."""
-        self._subscribed_entity_unique_identifiers.add(unique_identifier)
+        self._subscribed_entity_unique_identifiers[unique_identifier] = custom_identifier
 
     def remove_subscribed_entity_unique_identifier(self, unique_identifier: str) -> None:
         """Remove a program button."""
         if unique_identifier in self._subscribed_entity_unique_identifiers:
-            self._subscribed_entity_unique_identifiers.remove(unique_identifier)
+            del self._subscribed_entity_unique_identifiers[unique_identifier]
 
     @property
     def version(self) -> str | None:
@@ -607,21 +607,43 @@ class CentralUnit:
         d_address = get_device_address(address=address)
         return self._devices.get(d_address)
 
+    def get_entities(
+        self, registered_only: bool = False
+    ) -> Mapping[HmPlatform, Set[CallbackEntity]]:
+        """Return all externally registered entities."""
+        all_entities: dict[HmPlatform, set[CallbackEntity]] = {}
+        for platform in HmPlatform:
+            all_entities[platform] = set()
+        for device in self._devices.values():
+            for platform, entities in device.get_entities(registered_only=registered_only).items():
+                all_entities[platform].update(entities)
+
+        def add_to_dict(entities: Mapping[str, CallbackEntity]) -> None:
+            for entity in entities.values():
+                if (
+                    registered_only and entity.is_registered_externally
+                ) or registered_only is False:
+                    all_entities[entity.platform].add(entity)
+
+        add_to_dict(entities=self._sysvar_entities)
+        add_to_dict(entities=self._program_buttons)
+
+        return all_entities
+
     def get_entities_by_platform(
         self, platform: HmPlatform, exclude_subscribed: bool | None = None
     ) -> tuple[BaseEntity, ...]:
         """Return all entities by platform."""
-        existing_unique_ids = (
-            self._subscribed_entity_unique_identifiers if exclude_subscribed else ()
-        )
-
         return tuple(
             be
             for be in self._entities.values()
             if (
-                be.unique_identifier not in existing_unique_ids
-                and be.usage != EntityUsage.NO_CREATE
+                be.usage != EntityUsage.NO_CREATE
                 and be.platform == platform
+                and (
+                    not exclude_subscribed
+                    or (exclude_subscribed and be.is_registered_externally is False)
+                )
             )
         )
 
@@ -654,41 +676,39 @@ class CentralUnit:
         self, platform: HmPlatform, exclude_subscribed: bool | None = None
     ) -> tuple[GenericHubEntity, ...]:
         """Return the hub entities by platform."""
-        existing_unique_ids = (
-            self._subscribed_entity_unique_identifiers if exclude_subscribed else ()
-        )
-
         return tuple(
             he
             for he in (self.program_buttons + self.sysvar_entities)
-            if (he.unique_identifier not in existing_unique_ids and he.platform == platform)
+            if he.platform == platform
+            and (
+                not exclude_subscribed
+                or (exclude_subscribed and he.is_registered_externally is False)
+            )
+            and he.platform == platform
         )
 
     def get_update_entities(self, exclude_subscribed: bool | None = None) -> tuple[HmUpdate, ...]:
         """Return the update entities."""
-        existing_unique_ids = (
-            self._subscribed_entity_unique_identifiers if exclude_subscribed else ()
-        )
-
         return tuple(
             device.update_entity
             for device in self.devices
             if device.update_entity
-            and device.update_entity.unique_identifier not in existing_unique_ids
+            and (
+                not exclude_subscribed
+                or (exclude_subscribed and device.update_entity.is_registered_externally is False)
+            )
         )
 
     def get_channel_events_by_event_type(
         self, event_type: EventType, exclude_subscribed: bool | None = None
     ) -> tuple[list[GenericEvent], ...]:
         """Return all channel event entities."""
-        existing_unique_ids = (
-            self._subscribed_entity_unique_identifiers if exclude_subscribed else ()
-        )
-
         hm_channel_events: list[list[GenericEvent]] = []
         for device in self.devices:
             for channel_events in device.get_channel_events(event_type=event_type).values():
-                if channel_events[0].channel_unique_identifier not in existing_unique_ids:
+                if not exclude_subscribed or (
+                    exclude_subscribed and channel_events[0].is_registered_externally is False
+                ):
                     hm_channel_events.append(channel_events)
                     continue