+++ hahomematic/platforms/update.py
@@ -8,7 +8,8 @@ from __future__ import annotations
 from collections.abc import Callable
 from typing import Final
 
-from hahomematic.const import CallBackSource, HmPlatform
+from hahomematic.const import HmPlatform
+from hahomematic.exceptions import HaHomematicException
 from hahomematic.platforms import device as hmd
 from hahomematic.platforms.decorators import config_property, value_property
 from hahomematic.platforms.entity import CallbackEntity
@@ -70,21 +71,24 @@ class HmUpdate(CallbackEntity):
         return self._device.firmware_update_state
 
     def register_update_callback(
-        self, update_callback: Callable, source: CallBackSource = CallBackSource.EXTERNAL
+        self, update_callback: Callable, custom_identifier: str | None = None
     ) -> None:
         """Register update callback."""
         self._device.register_firmware_update_callback(update_callback)
-        if source == CallBackSource.EXTERNAL:
-            self._central.add_subscribed_entity_unique_identifier(
-                unique_identifier=self.unique_identifier
-            )
+        if custom_identifier is not None:
+            if self._custom_identifier is not None:
+                raise HaHomematicException(
+                    f"REGISTER_UPDATE_CALLBACK failed: hm_entity: {self.full_name} is already registered by {self._custom_identifier}"
+                )
+            self._custom_identifier = custom_identifier
 
     def unregister_update_callback(
-        self, update_callback: Callable, source: CallBackSource = CallBackSource.EXTERNAL
+        self, update_callback: Callable, custom_identifier: str | None = None
     ) -> None:
         """Unregister update callback."""
         self._device.unregister_firmware_update_callback(update_callback)
-        if source == CallBackSource.EXTERNAL:
+        if custom_identifier is not None:
+            self._custom_identifier = None
             self._central.remove_subscribed_entity_unique_identifier(
                 unique_identifier=self.unique_identifier
             )