+++ changelog.md
@@ -3,6 +3,7 @@
 - Adjust typing after move to more enums
 - Add measure_execution_time to writing methodes
 - Fix send sysvar #1249
+- Add HmIPW-SCTHD
 
 # Version 2023.10.6 (2023-10-08)